Better Auth
The email and domain you use is now to be hidden in the .env file with the API key.
All variables required to be changed to suit your setup is now all in the .env file. You should not need to change anything in the server.js or html file.