Added triggerScript.js
triggerScript.js is a file that will be used to bypass the need for the html file.
The html file was good for manual input and testing, but is unnecessary for the integration process.
Currently this file has pre-filled in data for testing purposes, but the idea is it uses placeholder variables that will be automatically filled by ManageEngine when a ticket is created and then sent off to Jira automatically.