Updated Input flexibility
The current server.js file did not account for no technician being assigned on the ServiceDesk Plus side. When no technician is specifically assigned SD+ uses the value "Not Specified". Server.js now takes this value and converts it to "Unassigned" so Jira can read it.
If needed, an automation rule can be set up so if a ticket comes in that is unassigned it can be assigned a technician automatically.