Added department
Created a new custom field for Department, which will take the department from ManageEngine and put it straight into this new field in Jira.
Attempted to create a way to check if the reporter is already a customer and if they are in an organisation, and if they are not then it will create a customer profile/create a new organisation based on that department. This did not work, and i settled for an automation rule that checks if the organisation rule is empty, and if it is, set it to the domain of the user's email. If there is no organisation with this name then it creates a new one.