Added Created By Field
Realised some tickets might not be made by the person who is requesting the service/info / reporting an incident. So if a person creates a ticket on behalf of someone it will be reflected in this field. Thought it would be good for auditing in the future in case any issues arise