Added category and subcategory fields
Added the category and subcategory fields so we can use them to detect what kind of ticket it is and specify a better issue/request type that fits the issue more.
For example, for account terminations in manageEngine the category is User administration and the subcategory is Employee Cessation. In Jira it will then change the issue type to User Admin and the request type to Account Terminations.