Added ID Field
Created a new custom field in order to keep track of the ticket number from ServiceDesk Plus in case anything needs to be cross-references or to make sure the ticket is syncing up correctly.

Also realised in the Account Cessation template the category is called Account Administrator rather than User Administrator, so added that to my automation rule in Jira to look for that too.